requirements.txt
ipython==8.15.0
ipywidgets==8.1.1
matplotlib==3.8.0
numpy==1.23.5
opencv_python==4.8.0.76
openvino==2023.1.0
openvino_dev==2023.1.0
Pillow==10.0.1
Pillow==10.1.0
Requests==2.31.0
torch==2.0.1
tqdm==4.66.1
ultralytics==8.0.181
